Refactor draggable preferences
